+++ apps/chatbot/scripts/run.test.sh
@@ -7,4 +7,4 @@ echo '-=-=-=-=-=-=-= create redis index =-=-=-=-=-=-=-=-'
 ./scripts/create_redis_index.sh
 
 echo '-=-=-=-=-=-=-=-=-=- run pytest -=-==-=-=-=-=-=-=-=-'
-pytest -vv -p no:warnings
+poetry run pytest src/app/
